introduce build watcher class
move  parts out of spaceship - so we have access to UI
