+++ fastlane_core/lib/fastlane_core.rb
@@ -30,6 +30,7 @@ require 'fastlane_core/tool_collector'
 require 'fastlane_core/fastlane_folder'
 require 'fastlane_core/keychain_importer'
 require 'fastlane_core/swag'
+require 'fastlane_core/build_watcher'
 
 # Third Party code
 require 'colored'
